+++ playground/src/components/details/editor/file_editor.tsx
