+++ playground/src/components/details/exercise_detail/programming.tsx
@@ -1,19 +1,19 @@
 import type { ProgrammingExercise } from "@/model/exercise";
 
 import Disclosure from "@/components/disclosure";
-import CodeView from "@/components/details/code_view";
+import CodeEditor from "@/components/details/editor/code_editor";
 
 export default function ProgrammingExerciseDetail({ exercise }: { exercise: ProgrammingExercise; }) {
   return (
     <>
       <Disclosure title="Template Repository">
-        <CodeView repository_url={exercise.template_repository_url} />
+        <CodeEditor repository_url={exercise.template_repository_url} />
       </Disclosure>
       <Disclosure title="Solution Repository">
-        <CodeView repository_url={exercise.solution_repository_url} />
+        <CodeEditor repository_url={exercise.solution_repository_url} />
       </Disclosure>
       <Disclosure title="Tests Repository">
-        <CodeView repository_url={exercise.tests_repository_url} />
+        <CodeEditor repository_url={exercise.tests_repository_url} />
       </Disclosure>
     </>
   );
